Add files via upload
Content
This dataset contains all the listed movies and TV shows on Netflix as of 2021. The dataset was collected from Flixable, a third-party Netflix search engine.

Acknowledgements
Firstly, I would like to thank Shivamb as a similar dataset for Netflix content was available. There were two shortcomings with the previously available database. Firstly, it was of the year 2019. Secondly, (which is the more important part), IMDb rating for the movies and TV shows was missing. This is a very important feature, especially when it comes to building recommendation algorithms.

Inspiration
This dataset can be used to perform the following:

EDA (exploratory data analysis)
Data visualization
Building a Recommendation Algorithm